Title: palindrome 여부 판단하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240719/palindrome 여부 판단하기/determine-whether-palindrome-is-present.cpp
@@ -6,8 +6,8 @@ using namespace std;
 string str;
 
 bool Palindrome(string &s) {
-    for(int i = 0; i < (int) s.length(); i++)
-        if(s[i] != s[(int) s.length() - i - 1])
+    for(int i = 0; i < s.length(); i++)
+        if(s[i] != s[ s.length() - i - 1])
             return false;
     
     return true;
